Support storage HMAC key Issue: https://github.com/upbound/provider-gcp/issues/51 This resource is needed to do S3-style access on GCS buckets
As this comment shows: https://github.com/upbound/provider-gcp/issues/197#issuecomment-1413437898, this resource causes an import cycle
However, we can resolve that by putting the resource in the `cloudplatform` package.
While, this key is specific to storage, it is still just a product of a service account and isn't linked to any storage resources.

Signed-off-by: Julien Duchesne <julien.duchesne@grafana.com>