Ignore node_count and initial_node_count when late initializing NodePools
Having node_count late-initialized with autoscaling can cause the autoscaler and
Crossplane to fight over how many nodes there should be (with Crossplane
effectively reverting autoscaler changes). If autoscaling is used, node_count
should remain unset.

As documented in [0], initial_node_count may (or may not?) change if the node
count is manually changed (via gcloud or the Cloud Console). This can cause
resources to fail reconciliation because terraform wants to destroy and recreate
the node pool. If the node count is changed manually, either Crossplane should
correct it back to the value of an explicitly set node_count field, or it should
be ignored because autoscaler is in use.

[0]: https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/container_node_pool#initial_node_count

Signed-off-by: Samuel Littley <slittley@thoughtmachine.net>